BankAccount.java
Review
What’s the use of having variables if you can’t do anything with them? We’ve now seen some ways you can operate on variables and compare them. The possibilities are endless!

We covered:

Addition and subtraction, using + and -
Multiplication and division, using * and /
The modulo operator for finding remainders, %
Compound assignment operators +=, -=, *=, /=, and %=.
The order of operations: parentheses -> exponents -> multiplication, division, modulo -> addition, subtraction
Greater than, >, and less than, <
Equal to, ==, and not equal to, !=
Greater than or equal to, >=, and less than or equal to, <=
equals() for comparing Strings and other objects
Using + to concatenate Strings
The final keyword which makes variables unchangeable
Practice some of these concepts here, to make sure you have a solid foundation for learning more complicated and exciting Java concepts!